+++ src/main/java/homeTry/auth/kakaoAuth/dto/TokenResponseDTO.java
@@ -1,4 +1,4 @@
-package homeTry.dto.kakaoAPI;
+package homeTry.auth.kakaoAuth.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.constraints.NotNull;
